+++ tests/models.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from django.utils.encoding import python_2_unicode_compatible
 from django.db import models
+from django.utils.encoding import python_2_unicode_compatible
+from django.utils.translation import ugettext_lazy as _
 
 
 @python_2_unicode_compatible
@@ -40,7 +41,7 @@ COUNTRIES = {
 
 @python_2_unicode_compatible
 class Manufacturer(models.Model):
-    name = models.CharField(max_length=255)
+    name = models.CharField(max_length=255, default=_("Test lazy tanslation"))
     country_code = models.CharField(max_length=2)
     created = models.DateField()
     logo = models.ImageField(blank=True)